feat: add Windows enable mode
Currently we try to enable the auto launch dynamically (system-wide first if we have admin privileges, with a fallback to current user). In some cases it might make sense to configure this behavior though - for instance when the app is installed to the current user only, and you want to avoid the local machine registry.